+++ scr/compromise_game/engines.py
